# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Aplicativo Web para la Centralización de Documentos Electrónicos Tributarios",
        "contact": {
            "name": "Rafael Larrea",
            "email": "jrafael1108@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/empresa/list/{estado}": {
            "get": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Obtiene una lista de empresas por estado.",
                "operationId": "3def7937343e0598fa974ec530e7ce30",
                "parameters": [
                    {
                        "name": "estado",
                        "in": "path",
                        "description": "Filtra la empresa por estado.",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Token": []
                    },
                    {
                        "Session": []
                    }
                ]
            }
        },
        "/api/empresa/api_key": {
            "get": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Obtiene una lista de empresas por busqueda.",
                "operationId": "b552525ce27eb7ebb62085b40f58f74b",
                "responses": {
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Token": []
                    },
                    {
                        "Session": []
                    }
                ]
            }
        },
        "/api/empresa/buscar/{filtro}": {
            "get": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Obtiene una lista de empresas por busqueda.",
                "operationId": "883f87c4a0c05542daaaeb55fe390ac6",
                "parameters": [
                    {
                        "name": "filtro",
                        "in": "path",
                        "description": "Filtra la empresa por ruc, razon_social, nombre_comercial.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Token": []
                    },
                    {
                        "Session": []
                    }
                ]
            }
        },
        "/api/empresa/{id}": {
            "get": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Obtener detalles de una empresa por ID",
                "operationId": "0aae62a86426eb3f6c666da4bdae026c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la empresa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la empresa"
                    },
                    "404": {
                        "description": "Empresa no encontrada"
                    }
                },
                "security": [
                    {
                        "Token": []
                    },
                    {
                        "Session": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Actualiza la entidad empresa",
                "operationId": "5efb82a74a0398d20204102e2752899b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la empresa a actualizar.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "NombreComercial": {
                                        "type": "string"
                                    },
                                    "PuedeEnviarCorreo": {
                                        "type": "boolean",
                                        "example": 0
                                    },
                                    "UsaConfiguracionSMTP": {
                                        "type": "boolean",
                                        "example": 0
                                    },
                                    "Email": {
                                        "type": "string"
                                    },
                                    "Servidor": {
                                        "type": "string"
                                    },
                                    "Puerto": {
                                        "type": "integer"
                                    },
                                    "TipoSeguridad": {
                                        "type": "integer",
                                        "example": 0
                                    },
                                    "UsuarioSMTP": {
                                        "type": "string"
                                    },
                                    "PasswordSMTP": {
                                        "type": "string"
                                    },
                                    "APIKey": {
                                        "type": "string"
                                    },
                                    "Logo": {
                                        "type": "string"
                                    },
                                    "IdUsuario": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Token": []
                    },
                    {
                        "Session": []
                    }
                ]
            }
        },
        "/api/empresa/": {
            "post": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Crear una nueva empresa",
                "operationId": "0dc2d8c3fac74dce0b488b585dd226a5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "RUC": {
                                        "type": "string"
                                    },
                                    "RazonSocial": {
                                        "type": "string"
                                    },
                                    "NombreComercial": {
                                        "type": "string"
                                    },
                                    "PuedeEnviarCorreo": {
                                        "type": "boolean",
                                        "example": 0
                                    },
                                    "UsaConfiguracionSMTP": {
                                        "type": "boolean",
                                        "example": 0
                                    },
                                    "Email": {
                                        "type": "string"
                                    },
                                    "Servidor": {
                                        "type": "string"
                                    },
                                    "Puerto": {
                                        "type": "integer"
                                    },
                                    "TipoSeguridad": {
                                        "type": "integer",
                                        "example": 0
                                    },
                                    "UsuarioSMTP": {
                                        "type": "string"
                                    },
                                    "PasswordSMTP": {
                                        "type": "string"
                                    },
                                    "APIKey": {
                                        "type": "string"
                                    },
                                    "Logo": {
                                        "type": "string"
                                    },
                                    "IdUsuario": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Token": []
                    },
                    {
                        "Session": []
                    }
                ]
            }
        },
        "/api/empresa/deshabilitar/{id}": {
            "post": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Deshabilita una entidad empresa",
                "operationId": "d2b7500ab441caa7efc997146f573cd8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la empresa a deshabilitar.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Token": []
                    },
                    {
                        "Session": []
                    }
                ]
            }
        },
        "/api/empresa/habilitar/{id}": {
            "post": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Habilita una entidad empresa",
                "operationId": "49545ffe9d06f9d3295e3f26474ed910",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la empresa a habilitar.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Token": []
                    },
                    {
                        "Session": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "APIResponse": {
                "description": "",
                "properties": {
                    "Success": {
                        "type": "boolean"
                    },
                    "Message": {
                        "type": "string"
                    },
                    "Data": {
                        "type": "object"
                    },
                    "Exception": {
                        "type": "object"
                    },
                    "Code": {
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "Token": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "bearerFormat": "JWT",
                "in": "header"
            },
            "Session": {
                "type": "apiKey",
                "description": "Enter your session id",
                "name": "SessionId",
                "in": "header"
            },
            "Apikey": {
                "type": "apiKey",
                "description": "Enter your APIKEY for company",
                "name": "APIKEY",
                "in": "header"
            }
        }
    },
...
}